+++ drake/systems/framework/state_vector.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <cstdint>
 #include <stdexcept>
 
 #include <Eigen/Dense>
@@ -23,21 +24,21 @@ class StateVector {
   ///
   /// Implementations should ensure this operation is O(1) and allocates no
   /// memory.
-  virtual ptrdiff_t size() const = 0;
+  virtual int64_t size() const = 0;
 
   /// Returns the element at the given index in the vector. Throws
   /// std::out_of_range if the index is >= size().
   ///
   /// Implementations should ensure this operation is O(1) and allocates no
   /// memory.
-  virtual const T GetAtIndex(ptrdiff_t index) const = 0;
+  virtual const T GetAtIndex(int64_t index) const = 0;
 
   /// Replaces the state at the given index with the value. Throws
   /// std::out_of_range if the index is >= size().
   ///
   /// Implementations should ensure this operation is O(1) and allocates no
   /// memory.
-  virtual void SetAtIndex(ptrdiff_t index, const T& value) = 0;
+  virtual void SetAtIndex(int64_t index, const T& value) = 0;
 
   /// Replaces the entire state with the contents of value. Throws
   /// std::out_of_range if value is not a column vector with size() rows.
@@ -62,7 +63,7 @@ class StateVector {
     if (vec.rows() != size()) {
       throw std::out_of_range("Addends must be the same length.");
     }
-    for (ptrdiff_t i = 0; i < size(); ++i) {
+    for (int64_t i = 0; i < size(); ++i) {
       vec[i] += GetAtIndex(i);
     }
   }
@@ -78,7 +79,7 @@ class StateVector {
       throw std::out_of_range("Addends must be the same length.");
     }
     assert(size() == rhs.size());
-    for (ptrdiff_t i = 0; i < size(); ++i) {
+    for (int64_t i = 0; i < size(); ++i) {
       SetAtIndex(i, GetAtIndex(i) + rhs.GetAtIndex(i));
     }
     return *this;
